+++ crates/task-expander/src/lib.rs
@@ -1,9 +1,11 @@
 mod task_expander;
 mod task_expander_error;
+mod task_lookup;
 mod token_expander;
 mod token_expander_error;
 
 pub use task_expander::*;
 pub use task_expander_error::*;
+pub use task_lookup::*;
 pub use token_expander::*;
 pub use token_expander_error::*;